+++ basicalter/slice_test.go
@@ -29,7 +29,7 @@ func TestDelEmptyStrings(t *testing.T) {
 
 	if v := basicalter.DelEmptyStrings(sliceOfString); len(v) != 2 {
 		t.Errorf("DelEmptyStrings didn't remove empty string: %v", v)
-	} else if !basiccheck.EqualStringSlice(v, []string{"foo", "bar"}) {
+	} else if !basiccheck.EqualSlice(v, []string{"foo", "bar"}) {
 		t.Errorf("DelEmptyStrings didn't remove empty string: %v", v)
 	}
 }
@@ -39,7 +39,7 @@ func TestDelStringInSlice(t *testing.T) {
 
 	if v := basicalter.DelStringInSlice("baz", sliceOfString); len(v) != 2 {
 		t.Errorf("DelStringInSlice didn't remove 'baz': %v", v)
-	} else if !basiccheck.EqualStringSlice(v, []string{"foo", "bar"}) {
+	} else if !basiccheck.EqualSlice(v, []string{"foo", "bar"}) {
 		t.Errorf("DelStringInSlice didn't remove 'baz': %v", v)
 	}
 }
@@ -51,7 +51,7 @@ func TestFilterStringsWith(t *testing.T) {
 		return strings.HasPrefix(s, "ba")
 	}); len(v) != 3 {
 		t.Errorf("FilterStringsWith didn't remove foo (without prefix 'ba'): %v", v)
-	} else if !basiccheck.EqualStringSlice(v, []string{"baz", "bar", "baz"}) {
+	} else if !basiccheck.EqualSlice(v, []string{"baz", "bar", "baz"}) {
 		t.Errorf("FilterStringsWith didn't remove foo (without prefix 'ba'): %v", v)
 	}
 }
@@ -62,7 +62,7 @@ func TestReverseSlice(t *testing.T) {
 	basicalter.ReverseStrings(sliceOfString)
 
 	desiredSlice := []string{"Hello", "World", "bar", "baz", "foo"}
-	if !basiccheck.EqualStringSlice(sliceOfString, desiredSlice) {
+	if !basiccheck.EqualSlice(sliceOfString, desiredSlice) {
 		t.Errorf("ReverseStrings didn't reverse slice: %v expected %v", sliceOfString, desiredSlice)
 	}
 }
@@ -73,7 +73,7 @@ func TestSortStringsByLengthInc(t *testing.T) {
 	basicalter.SortStringsByLengthInc(s)
 
 	desiredSlice := []string{"Go", "Grin", "Alpha", "Bravo", "Delta", "Gopher"}
-	if !basiccheck.EqualStringSlice(s, desiredSlice) {
+	if !basiccheck.EqualSlice(s, desiredSlice) {
 		t.Errorf("SortStringsByLength didn't sort slice with smaller first and lexicographic order")
 	}
 }
@@ -84,7 +84,7 @@ func TestSortStringsByLengthDec(t *testing.T) {
 	basicalter.SortStringsByLengthDec(s)
 
 	desiredSlice := []string{"Gopher", "Alpha", "Bravo", "Delta", "Grin", "Go"}
-	if !basiccheck.EqualStringSlice(s, desiredSlice) {
+	if !basiccheck.EqualSlice(s, desiredSlice) {
 		t.Errorf("SortStringsByLength didn't sort slice with smaller last and lexicographic order")
 	}
 }
@@ -94,7 +94,7 @@ func TestReplaceStringsWith(t *testing.T) {
 
 	basicalter.ReplaceStringsWith(sliceOfString, strings.ToLower)
 
-	if !basiccheck.EqualStringSlice(sliceOfString, []string{"foo", "bar", "baz"}) {
+	if !basiccheck.EqualSlice(sliceOfString, []string{"foo", "bar", "baz"}) {
 		t.Errorf("ReplaceStringsWith didn't replace all strings in slice "+
 			"with the lowercase version: %v", sliceOfString)
 	}